feat: adicionando rotina para atualizar preços dos quartos com base na ultima cotação do dólar

+++ ReservaHotel.Domain/Model/DTOs/Quarto/AddUpdateQuartoDTO.cs
@@ -16,7 +16,7 @@ namespace ReservaHotel.Domain.Model.DTOs.Quarto
         public string Ocupacao { get; set; }
         public Guid HotelId { get; set; }
         public TipoQuarto? TipoQuarto { get; set; }
-        public float? DiariaDolar { get; set; }
+        public decimal? DiariaDolar { get; set; }
         public Guid Id { get; set; }
     }
 }
